requirements.txt
flask
transformers
textstat
python-docx
PyPDF2
sentencepiece